pyproject.toml
[tool.poetry]
name = "2026-main"
version = "0.1.0"
description = ""
authors = ["Gambhiro <gambhiro@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
splendidmoons = "^1.0.2"
doit = "^0.36.0"
mako = "^1.3.6"
pillow = "^11.0.0"
pymupdf = "^1.24.14"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"

[tool.ruff.lint]
ignore = ["E401", "E501", "E712"]
target-version = "py312"
